initial checkin of current version, prior to port to Python3

+++ clustermanager.py
@@ -4,6 +4,7 @@ import sys, os.path, tempfile, time, re, os
 import ConfigParser
 import subprocess
 import lockfile
+from dfcollection import DatafileCollection
 
 import requests
 from requests.packages.urllib3.exceptions import InsecureRequestWarning, InsecurePlatformWarning, SNIMissingWarning
@@ -50,7 +51,7 @@ class Cluster(object):
 	assert(keypath)
         if self._sshsession != None:
             self.ssh_session_end()
-        self._sshsession = os.path.abspath(".ubuntu@%(ip)s:22"%dict(ip=publicip))
+        self._sshsession = os.path.abspath(".ubuntu@%(ip)s:22.%(pid)s"%dict(ip=publicip,pid=os.getpid()))
 	cmd = self.ssh_session_start_cmd%dict(ip=publicip,keypath=keypath,sshsession=self._sshsession)
         if self._sshdebug:
             print >>sys.stderr, "Execute: "+cmd
@@ -87,7 +88,7 @@ class Cluster(object):
         else:
             return self.execute_(cmd)
 	
-    execcmd = "ssh -o \"ControlMaster=no\" -o \"ControlPath=%(sshsession)s\" ubuntu@%(ip)s \"%(cmd)s\""
+    execcmd = "ssh -o ConnectTimeout=10 -o \"ControlMaster=no\" -o \"ControlPath=%(sshsession)s\" ubuntu@%(ip)s \"%(cmd)s\""
     def execute_(self,cmd):
 	publicip = self.get('publicip')
 	assert(publicip)
@@ -163,11 +164,77 @@ class Cluster(object):
         self.execute("echo \"linux-x86_64\" > cptac-galaxy/PLATFORM")
         self.execute("( cd cptac-galaxy; sh ./update.sh )")
         self.copy(os.path.join(scriptdir,".galaxy.ini"),"cptac-galaxy/.galaxy.ini")
+	self.execute("chmod a+r cptac-galaxy/.galaxy.ini")
+	if os.path.exists(os.path.join(scriptdir,".rclone.conf")):
+	    self.copy(os.path.join(scriptdir,".rclone.conf"),"cptac-galaxy/.rclone.conf")
+	    self.execute("chmod a+r cptac-galaxy/.rclone.conf")
         self.execute("rm -rf cptac-galaxy/cptacdcc")
-        self.execute("wget -q -O - http://edwardslab.bmcb.georgetown.edu/software/downloads/CPTAC-DCC/cptacdcc.linux-x86_64.tgz | tar xzf - -C cptac-galaxy")
+        self.execute("wget --no-check-certificate -q -O - https://edwardslab.bmcb.georgetown.edu/software/downloads/CPTAC-DCC/cptacdcc.linux-x86_64.tgz | tar xzf - -C cptac-galaxy")
         self.dcccredentials()
         self.execute("mkdir -p " + self.datadir)
         self.ssh_session_end()
+
+    def version(self):
+	# local VERSION number
+	versions = dict()
+	versions['local_cptac_galaxy'] = open(os.path.join(os.path.split(os.path.abspath(sys.argv[0]))[0],'VERSION')).read().strip()
+	self.ssh_session_start()
+	for line in self.execute("cat /home/ubuntu/cptac-galaxy/VERSION",output=True):
+	    versions['remote_cptac_galaxy'] = line.strip()
+	    break
+	for line in self.execute("cat /home/ubuntu/galaxy-tools/extratools/VERSION",output=True):
+            versions['cptac_galaxy_tools'] = line.strip().split('-')[2]
+            break
+	for line in self.execute("cat /home/ubuntu/galaxy-tools/extratools/lib/CPTAC-CDAP-Reports/VERSION.txt",output=True):
+            versions['cdap_reports'] = line.strip()
+            break
+	for line in self.execute("/home/ubuntu/galaxy-tools/extratools/lib/cptac3-cdap/cptac-dcc/cptacdcc/cksum.sh --version",output=True):
+            versions['cptac_dcc_tools'] = line.strip().split()[2]
+            break
+	for line in self.execute(r"/home/ubuntu/galaxy-tools/extratools/lib/cptac3-cdap/cptac-mzid/cptacmzid/version.sh",output=True):
+            versions['cptac_dcc_mzidentml'] = line.strip()
+            break
+        self.ssh_session_end()
+	versions['cptac_galaxy_workflows'] = versions['remote_cptac_galaxy']
+	return versions
+
+    def update(self,version=None,tools_version=None):
+	self.ssh_session_start()
+	if not version:
+	    version = open(os.path.join(os.path.split(os.path.abspath(sys.argv[0]))[0],'VERSION')).read().strip()
+	print "Update cptac-galaxy and cptac-galaxy workflows to version %s ..."%(version,),
+	sys.stdout.flush()
+	for line in self.execute("( cd cptac-galaxy; sh ./update.sh %s )"%(version,),output=True):
+	    pass # print line
+	print "done."
+	sys.stdout.flush()
+	if not tools_version:
+	    for line in self.execute("cat cptac-galaxy/.defaults.ini",output=True):
+	        sl = line.split()
+	        if sl[0] == "tools_version":
+	            tools_version = sl[2].split('-',2)[-1]
+	assert tools_version
+	print "Update cptac-galaxy tools to %s ..."%(tools_version,),
+	sys.stdout.flush()
+	for line in self.execute("sudo -H -u galaxy sed -i 's/CPTAC-GALAXY-[0-9][0-9]*\.[0-9][0-9]*\.[0-9][0-9]*/CPTAC-GALAXY-%s/' /home/galaxy/.urls"%(tools_version,),output=True):
+	    pass # print line
+	for line in self.execute("( cd galaxy-tools/extratools; sh ./update.sh )",output=True):
+	    pass # print line
+	print "done."
+	sys.stdout.flush()
+	print "Update Galaxy workflows with cptac-galaxy workflows version %s ..."%(version,),
+        sys.stdout.flush()
+	for line in self.execute("./cptac-galaxy/wfupload --galaxy https://127.0.0.1/ --apikey %s --directory cptac-galaxy/workflows"%(self.get('apikey'),),output=True):
+	    pass # print line
+	print "done."
+	sys.stdout.flush()
+	print "Restart Galaxy ...",
+        sys.stdout.flush()
+	for line in self.execute("( cd galaxy-app; sudo -H -u galaxy ./rolling_restart.sh )",output=True):
+	    pass # print line
+	print "done."
+	sys.stdout.flush()
+	self.ssh_session_end()
 
     def jobdir(self,jobid):
         return "%s/%s"%(self.remotebase,jobid)
@@ -230,8 +297,81 @@ echo \"$!\" > execute.pid
         stopscript = ("""
 #!/bin/sh
 cd %(remotedir)s
-kill -9 `cat execute.pid`
+kill -9 `ps -ef | fgrep -w %(jobid)s | fgrep cptac-galaxy/execute | fgrep -v -w stop.sh | awk '{print $2}'`
         """%args).strip()
+
+        f,fn = tempfile.mkstemp()
+        os.close(f)
+        fh = open(fn,'w')
+        fh.write(stopscript)
+        fh.close()
+        self.copy(fn,'%s/stop.sh'%(rdir,))
+        os.unlink(fn)
+
+        self.ssh_session_end()
+
+        return jobid
+        
+    def setup_script_job(self,*args,**kwargs):
+        jobid = kwargs.get('jobid')
+        uploads = kwargs.get('uploads',[])
+        if 'jobid' in kwargs:
+            del kwargs['jobid']
+        if 'uploads' in kwargs:
+            del kwargs['uploads']
+        if not jobid:
+            import datetime
+            ts = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
+            jobid = "scriptjob-%s"%(ts,)
+        else:
+            assert not re.search(r'[^a-zA-Z0-9_.-]',jobid)
+        assert jobid not in self.list_jobids()
+        rdir = self.jobdir(jobid)
+        self.ssh_session_start()
+        self.execute('( mkdir -p %s; touch %s/execute.pid )'%(rdir,rdir))
+        self.execute('mkdir -p %s/files'%(rdir,))
+        self.execute('mkdir -p %s/results'%(rdir,))
+        self.copy(".galaxy.ini",rdir + "/.galaxy.ini")
+        self.execute("sed -i 's/%s/127.0.0.1/' "%(self.get('publicip'),)+ rdir + "/.galaxy.ini")
+
+        for f in uploads:
+            ff = os.path.split(f)[1]
+            t = "/".join([rdir,'files',ff])
+            self.copy(f,t)
+        jobargs = {}
+        jobargs['jobid'] = jobid
+        jobargs['remotedir'] = rdir
+        jobargs['name'] = self.get('name')
+        jobargs['outdir'] = '%s/results'%(rdir,)
+
+        startscript = ("""
+#!/bin/sh
+cd %s
+/home/ubuntu/cptac-galaxy/%s \\
+        """%(rdir,args[0])).strip()+"\n"
+        for a in args[1:]:
+            startscript += "   %s \\\n"%(a,)
+        for k,v in kwargs.items():
+            startscript += "    --%s %s \\\n"%(a, kwargs.get(a),)
+        startscript = startscript%jobargs
+        startscript += ("""
+    >execute.log 2>&1 &
+echo \"$!\" > execute.pid
+        """).strip()
+        
+        f,fn = tempfile.mkstemp()
+        os.close(f)
+        fh = open(fn,'w')
+        fh.write(startscript)
+        fh.close()
+        self.copy(fn,'%s/execute.sh'%(rdir,))
+        os.unlink(fn)
+
+        stopscript = ("""
+#!/bin/sh
+cd %(remotedir)s
+kill -9 `ps -ef | fgrep -w %(jobid)s | fgrep cptac-galaxy/execute | fgrep -v -w stop.sh | awk '{print $2}'`
+        """%jobargs).strip()
 
         f,fn = tempfile.mkstemp()
         os.close(f)
@@ -323,7 +463,7 @@ kill -9 `cat execute.pid`
         for l in self.execute("cd %s; /home/ubuntu/cptac-galaxy/organize1all.sh %s *"%(rdir,jobid),output=True):
             print l
         destdir = self.datadir + "/" + jobid
-        self.execute("mkdir -p %s"%(destdir,))
+        self.execute("rm -rf %s; mkdir -p %s"%(destdir,destdir))
         self.execute("cd %s; mv -f mzML PSM.tsv mzIdentML SummaryReports %s"%(rdir,destdir))
         self.ssh_session_end()
 
@@ -336,6 +476,22 @@ kill -9 `cat execute.pid`
         self.execute("mkdir -p %s/%s"%(self.datadir,destname))
         self.ssh_session_end()
 	self.rsyncup(fromdir,destdir)
+
+    def upload_file(self,fromfile,jobid):
+	assert(os.path.isfile(fromfile))
+	filepath,filename = os.path.split(fromfile)
+	destdir = self.jobdir(jobid) + "/files"
+	destfile = destdir + "/" + filename
+        self.ssh_session_start()
+	self.copy(fromfile,destfile)
+        self.ssh_session_end()
+
+    def list_files(self,jobid):
+	destdir = self.jobdir(jobid) + "/files"
+        self.ssh_session_start()
+	for line in self.execute("ls -l %s"%(destdir,),output=True):
+	    print line
+        self.ssh_session_end()
 
     def download_data(self,destdir,dataname,datatag):
         ddir = self.datadir + "/" + dataname + "/" + datatag;
@@ -369,15 +525,28 @@ kill -9 `cat execute.pid`
 		print "\t".join([l[:-6],"local",self.datadir + "/" + base + "/" + dir + "/" + l])
         self.ssh_session_end()
 
-    def status_job(self,jobid):
+    def status_job(self,jobid,all=False):
         try:
             self.ssh_session_start()
-            for line in self.execute("tail -n 30 -f %s/execute.log"%(self.jobdir(jobid),),output=True):
-                print >>sys.stderr, line
+            logs = "execute"
+	    if all:
+		logs = "*"
+            for line in self.execute("tail -n 30 -f %s/%s.log"%(self.jobdir(jobid),logs),output=True):
+                print line
         except KeyboardInterrupt:
             self.execute("ps -ef | fgrep %s/execute.log | fgrep tail | awk '{print \\$2}' | xargs -n 30 kill -9"%(jobid,))
         finally:
             self.ssh_session_end()
+
+    def job_logfile(self,jobid,filename="execute.log",tail=None):
+        self.ssh_session_start()
+	if tail != None:
+            for line in self.execute("tail -n %s %s/%s"%(tail,self.jobdir(jobid),filename),output=True):
+                print line
+	else:
+            for line in self.execute("cat %s/%s"%(self.jobdir(jobid),filename),output=True):
+                print line
+        self.ssh_session_end()
 
     def running_jobs(self,needsession=True):
         running = set()
@@ -385,28 +554,69 @@ kill -9 `cat execute.pid`
             if needsession:
                 self.ssh_session_start()
             jobsnames = self.list_jobids(needsession=False)
-            pids = set()
-            for line in self.execute("cat %s/*/execute.pid"%(self.remotebase,),output=True):
-                for w in line.split():
-                    pids.add(w)
-            for line in self.execute("ps %s"%(" ".join(pids)),output=True):
-                for w in line.split():
-                    if w in jobsnames:
-                        running.add(w)
+	    pid2job = {}
+	    for line in self.execute("grep '[0-9]' %s/*/execute.pid /dev/null"%(self.remotebase,),output=True):
+	        if line.startswith('grep:'):
+		    continue
+		sl = line.split(':',1)
+		job = sl[0].split('/')[-2]
+		pid = sl[1]
+	        pid2job[pid] = job
+	    for line in self.execute("ps %s"%(" ".join(pid2job)),output=True):
+		sl = line.split()
+		if sl[0] in pid2job:
+		    running.add(pid2job[sl[0]])
             return sorted(running)
         finally:
             if needsession:
                 self.ssh_session_end()
 
-    def status_all(self):
+    def status_all(self,all=False):
+        logs = "execute"
+        if all:
+	    logs = "*"
         try:
             self.ssh_session_start()
-            for line in self.execute("tail -n 30 -f %s/*/execute.log"%(self.remotebase,),output=True):
-                print >>sys.stderr, line
+            for line in self.execute("tail -n 30 -f %s/*/%s.log"%(self.remotebase,logs),output=True):
+                print line
         except KeyboardInterrupt:
-            self.execute("ps -ef | fgrep /execute.log | fgrep tail | awk '{print \\$2}' | xargs -n 30 kill -9")
+            self.execute("ps -ef | fgrep 'tail -n 30 -f ' | awk '{print \\$2}' | xargs -n 30 kill -9")
         finally:
             self.ssh_session_end()
+
+    def clear_workflow_instance(self,jobid,base):
+        gi = GalaxyInstance(url=self.get('url'),key=self.get('apikey'))
+        gi.verify=False
+        history = None
+        for h in gi.histories.get_histories(name=jobid):
+            history = h.get('id')
+	    break
+	if not history:
+	    return
+	base,extn = DatafileCollection.dssplit(base)
+	for di in gi.histories.show_matching_datasets(history,re.compile("^%s[-.]"%(re.escape(base),))):
+            id,name,state,visible,deleted = map(di.get,('id','name','state','visible','deleted'))
+	    print "Removing %s from Galaxy history %s."%(name,jobid)
+            gi.histories.delete_dataset(history,id)
+            gi.histories.delete_dataset(history,id,purge=True)
+
+    def clear_job(self,jobid):
+	self.ssh_session_start()
+	running = self.running_jobs(needsession=False)
+	isrunning = (jobid in running)
+        rdir = self.jobdir(jobid)
+	if isrunning:
+	    print "Stoping job: %s."%(jobid,)
+            self.execute("sh %s/stop.sh"%(rdir,))
+        gi = GalaxyInstance(url=self.get('url'),key=self.get('apikey'))
+        gi.verify=False
+        for h in gi.histories.get_histories(name=jobid):
+	    print "Removing Galaxy history %s."%(jobid,)
+            gi.histories.delete_history(h.get('id'), purge=True)
+	if isrunning:
+	    print "Restarting job: %s."%(jobid,)
+            self.execute("sh %s/execute.sh"%(rdir,))
+        self.ssh_session_end()
 
     def remove_job(self,jobid):
         gi = GalaxyInstance(url=self.get('url'),key=self.get('apikey'))
@@ -414,8 +624,9 @@ kill -9 `cat execute.pid`
         for h in gi.histories.get_histories(name=jobid):
             gi.histories.delete_history(h.get('id'), purge=True)
         rdir = self.jobdir(jobid)
+        ddir = self.datadir + "/" + jobid
         self.ssh_session_start()
-        self.execute("( sh %s/stop.sh; rm -rf %s )"%(rdir,rdir))
+        self.execute("( sh %s/stop.sh; rm -rf %s; rm -rf %s )"%(rdir,rdir,ddir))
         self.ssh_session_end()
 
     def remove_all_jobs(self):
@@ -435,7 +646,7 @@ class ClusterManager(object):
         self.config = ConfigParser.SafeConfigParser()
         try:
             lock.acquire()
-            found = self.config.read([self.iniFileGeneral,self.iniFile])
+            found = self.config.read([self.iniFileGeneral,os.path.join('cptac-galaxy',self.iniFile),self.iniFile])
         finally:
             lock.release()
         assert found > 0, "Can't find %s file(s)"%(self.iniFile,)
@@ -525,9 +736,13 @@ class ClusterManager(object):
         if not cluster:
             if name:
                 print >>sys.stderr, "Cluster \"%s\" not found.\n"%(name,)
-            print >>sys.stderr, "Clusters:"
-            for sec in self.getclusternames(type=type):
-                print >>sys.stderr, " ",sec
+	    names = self.getclusternames(type=type)
+	    if len(names) == 0:
+                print >>sys.stderr, "No clusters launched from this directory."
+	    else:
+                print >>sys.stderr, "Clusters:"
+                for sec in names:
+                    print >>sys.stderr, " ",sec
             sys.exit(1)
 	return cluster
